all methods are created

+++ app/src/main/java/Statistic/ApplianceModel.java
@@ -7,19 +7,19 @@ import java.util.Comparator;
  */
 public class ApplianceModel {
     private String appliname, date, frontelectusage, locationname;
-    private long elecusage, elecusageLocation;
+    private double elecusage, elecusageLocation;
     private boolean isLocation = false;
 
     public ApplianceModel() {
     }
 
-    public ApplianceModel(String appliname, long elecusage, String date, String frontelectusage) {
+    public ApplianceModel(String appliname, double elecusage, String date, String frontelectusage) {
         this.appliname = appliname;
         this.elecusage = elecusage;
         this.date = date;
         this.frontelectusage = frontelectusage;
     }
-    public ApplianceModel(String locationname, long elecusageLocation, String date, String frontelectusage,boolean isLocation){
+    public ApplianceModel(String locationname, double elecusageLocation, String date, String frontelectusage,boolean isLocation){
         this.locationname = locationname;
         this.elecusageLocation = elecusageLocation;
         this.date = date;
@@ -44,7 +44,7 @@ public class ApplianceModel {
         this.date = date;
     }
 
-    public long getElect() {
+    public double getElect() {
         return elecusage;
     }
 
@@ -52,15 +52,15 @@ public class ApplianceModel {
         return frontelectusage;
     }
 
-    public void setElect(long Elect) {
+    public void setElect(double Elect) {
         this.elecusage = Elect;
     }
 
     public static Comparator<ApplianceModel> electusageMaxComparator = new Comparator<ApplianceModel>() {
 
         public int compare(ApplianceModel e1, ApplianceModel e2) {
-            long electusage1 = e1.getElect();
-            long electusage2 = e2.getElect();
+            double electusage1 = e1.getElect();
+            double electusage2 = e2.getElect();
 
             //ascending order
             //return appliname1.compareTo(appliname2);
@@ -72,8 +72,8 @@ public class ApplianceModel {
     public static Comparator<ApplianceModel> electusageMinComparator = new Comparator<ApplianceModel>() {
 
         public int compare(ApplianceModel e1, ApplianceModel e2) {
-            long electusage1 = e1.getElect();
-            long electusage2 = e2.getElect();
+            double electusage1 = e1.getElect();
+            double electusage2 = e2.getElect();
 
             //ascending order
             return (int) (electusage1-electusage2);
